+++ src/module/data/actor/base.mjs
@@ -76,6 +76,9 @@ export default class BaseActorModel extends SubtypeModelMixin(foundry.abstract.T
 
     this.potency = {
       bonuses: 0,
+      weak: 0,
+      average: 0,
+      strong: 0,
     };
 
     this.statuses = {
@@ -100,6 +103,12 @@ export default class BaseActorModel extends SubtypeModelMixin(foundry.abstract.T
     super.prepareDerivedData();
 
     this.stamina.winded = Math.floor(this.stamina.max / 2);
+
+    const highestCharacteristic = Math.max(0, ...Object.values(this.characteristics).map(c => c.value));
+
+    this.potency.weak += highestCharacteristic - 2 + this.potency.bonuses;
+    this.potency.average += highestCharacteristic - 1 + this.potency.bonuses;
+    this.potency.strong += highestCharacteristic + this.potency.bonuses;
 
     // Set movement speeds when affected by grabbed, restrained, or slowed
     const isSlowed = this.parent.statuses.has("slowed");
@@ -182,7 +191,10 @@ export default class BaseActorModel extends SubtypeModelMixin(foundry.abstract.T
    * @param {import("@common/abstract/_types.mjs").DatabaseUpdateOperation} operation
    * @param {User} user
    */
-  _preUpdate(changes, operation, user) {
+  async _preUpdate(changes, options, user) {
+    const allowed = await super._preUpdate(changes, options, user);
+    if (allowed === false) return false;
+
     const newSize = foundry.utils.getProperty(changes, "system.combat.size.value");
     if ((newSize !== undefined) && (this.combat.size.value !== newSize)) {
       foundry.utils.mergeObject(changes, {