+++ src/module/data/pseudo-documents/power-roll-effects/_module.mjs
@@ -1,2 +1,5 @@
+export { default as AppliedPowerRollEffect } from "./applied-effect.mjs";
 export { default as BasePowerRollEffect } from "./base-power-roll-effect.mjs";
 export { default as DamagePowerRollEffect } from "./damage-effect.mjs";
+export { default as ForcedMovementPowerRollEffect } from "./forced-movement-effect.mjs";
+export { default as OtherPowerRollEffect } from "./other-effect.mjs";